pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "btc-alpha-engine"
version = "0.1.0"
description = "Adaptive BTC alpha engine with derivatives-based signals, walk-forward research, and paper/live trading on Hyperliquid"
authors = ["Alpha Team"]
license = "Proprietary"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5"
pydantic-settings = "^2.1"
sqlalchemy = {version = "^2.0", extras = ["asyncio"]}
alembic = "^1.13"
psycopg2-binary = {version = "^2.9", optional = true}
pandas = "^2.1"
polars = "^0.20"
numpy = "^1.26"
scipy = "^1.12"
scikit-learn = "^1.4"
lightgbm = "^4.2"
xgboost = "^2.0"
httpx = "^0.27"
websockets = "^12.0"
structlog = "^24.1"
apscheduler = "^3.10"
pyyaml = "^6.0"
python-dotenv = "^1.0"
joblib = "^1.3"
tenacity = "^8.2"
click = "^8.1"
prometheus-client = "^0.21"

[tool.poetry.extras]
postgres = ["psycopg2-binary"]
torch = ["torch"]
lium = ["lium.io"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^4.1"
pytest-asyncio = "^0.23"
mypy = "^1.8"
ruff = "^0.2"
black = "^24.1"
ipython = "^8.20"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
